refactor(core): move document trait for CodeSearchDocument to index/code/document.rs
update
