+++ crates/tabby-common/src/index/code/mod.rs
@@ -1,3 +1,5 @@
+mod document;
+
 use lazy_static::lazy_static;
 use tantivy::{
     query::{BooleanQuery, ConstScoreQuery, Query, TermQuery},
